Move BCH from QuantumClifford to QECCore
Define AbstractPolynomialCode before the Goppa and BCH includes, then place the BCH type and code_n in QECCore core. Keep generator_polynomial, parity_matrix, and code_k in the Nemo extension.

Move the existing BCH tests with the implementation and add a focused QuantumClifford compatibility test. The legacy QuantumClifford.ECC bindings remain the exact QECCore type objects, preserving dispatch, type annotations, isa checks, and serialized references.
