+++ lib/QECCore/ext/QECCoreNemoExt/goppa.jl
@@ -1,5 +1,3 @@
-abstract type AbstractPolynomialCode <: AbstractCECC end
-
 """
     $TYPEDEF
